Create CpioLruCache struct to hold all the lru related info
This makes it easier to `read()` the `RwLock` and access the max and
current sizes at the same time, without needing to open another `read()`
on another `RwLock`.

We have a `prune` function to clean up the LRU cache entry, as well as
a `push` function that will call `prune` if it would exceed the max
configured cache size.