Take a write lock instead of a read and then write lock
Otherwise, we could deadlock here because we're logging the sizes of the
cache using a read lock which doesn't release until it goes out of scope
(and `cache.write()` just ends up blocking forever).

We could just explicitly drop the read lock instead, since we're not
actually writing to the cache sizes, but... This makes the logic easier
to reason about.