init OpenedCpio struct to hold a handle to the CPIO file, fixup pruning
When setting a maximum cache size of 0, we should essentially clean up
every CPIO after we've acquired its handle. This way, the disk cache
doesn't exceed the maximum size, but we can still serve any clients who
needed that specific CPIO (until that CPIO has been served, and then the
handle will be dropped with the OpenedCpio).

Also got rid of the epsilon, since that would allow the cache to grow
up to that size in the case of a max cache size smaller than any CPIOs
(e.g. 0).